Workaround user rest

+++ src/main/java/com/project/workshopmanagment/repository/UserRepository.java
@@ -3,7 +3,8 @@ package com.project.workshopmanagment.repository;
 import com.project.workshopmanagment.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.repository.CrudRepository;
+import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
-public interface UserRepository extends CrudRepository<User, Integer> {
+public interface UserRepository extends CrudRepository<User, Long> {
     User findByEmail(String email);
 }
